fixed bug in typedef exception reporting
fixes #1167
